+++ code/client/core/recipe.sml
@@ -105,6 +105,26 @@ Tests: {
 		'tests/gen/build/recipe-build-location-manager-tests.gen.cpp'
 		'tests/gen/main.cpp'
 		'tests/gen/module.cpp'
+		'tests/gen/local-user-config/local-user-config-extensions-tests.gen.cpp'
+		'tests/gen/local-user-config/local-user-config-tests.gen.cpp'
+		'tests/gen/operation-graph/operation-graph-tests.gen.cpp'
+		'tests/gen/operation-graph/operation-graph-manager-tests.gen.cpp'
+		'tests/gen/operation-graph/operation-graph-reader-tests.gen.cpp'
+		'tests/gen/operation-graph/operation-graph-writer-tests.gen.cpp'
+		'tests/gen/operation-graph/operation-results-tests.gen.cpp'
+		'tests/gen/operation-graph/operation-results-manager-tests.gen.cpp'
+		'tests/gen/operation-graph/operation-results-reader-tests.gen.cpp'
+		'tests/gen/operation-graph/operation-results-writer-tests.gen.cpp'
+		'tests/local-user-config/local-user-config-extensions-tests.cpp'
+		'tests/local-user-config/local-user-config-tests.cpp'
+		'tests/operation-graph/operation-graph-tests.cpp'
+		'tests/operation-graph/operation-graph-manager-tests.cpp'
+		'tests/operation-graph/operation-graph-reader-tests.cpp'
+		'tests/operation-graph/operation-graph-writer-tests.cpp'
+		'tests/operation-graph/operation-results-tests.cpp'
+		'tests/operation-graph/operation-results-manager-tests.cpp'
+		'tests/operation-graph/operation-results-reader-tests.cpp'
+		'tests/operation-graph/operation-results-writer-tests.cpp'
 		'tests/utilities/test-helpers.cpp'
 	]
 	IncludePaths: [
