+++ code/client/core/tests/gen/module.cpp
@@ -1,6 +1,7 @@
 
 export module Soup.Core.Tests;
 
+// build
 export import :BuildEngineTests;
 export import :BuildEvaluateEngineTests;
 export import :BuildHistoryCheckerTests;
@@ -10,6 +11,21 @@ export import :FileSystemStateTests;
 export import :PackageProviderTests;
 export import :RecipeBuildLocationManagerTests;
 
+// local-user-config
+export import :LocalUserConfigExtensionsTests;
+export import :LocalUserConfigTests;
+
+// operation-graph
+export import :OperationGraphTests;
+export import :OperationGraphManagerTests;
+export import :OperationGraphReaderTests;
+export import :OperationGraphWriterTests;
+export import :OperationResultsTests;
+export import :OperationResultsManagerTests;
+export import :OperationResultsReaderTests;
+export import :OperationResultsWriterTests;
+
+// build
 export import :RunBuildEngineTests;
 export import :RunBuildEvaluateEngineTests;
 export import :RunBuildHistoryCheckerTests;
@@ -18,3 +34,17 @@ export import :RunBuildRunnerTests;
 export import :RunFileSystemStateTests;
 export import :RunPackageProviderTests;
 export import :RunRecipeBuildLocationManagerTests;
+
+// local-user-config
+export import :RunLocalUserConfigExtensionsTests;
+export import :RunLocalUserConfigTests;
+
+// operation-graph
+export import :RunOperationGraphTests;
+export import :RunOperationGraphManagerTests;
+export import :RunOperationGraphReaderTests;
+export import :RunOperationGraphWriterTests;
+export import :RunOperationResultsTests;
+export import :RunOperationResultsManagerTests;
+export import :RunOperationResultsReaderTests;
+export import :RunOperationResultsWriterTests;
