+++ code/client/core/tests/local-user-config/local-user-config-extensions-tests.cpp
@@ -1,12 +1,32 @@
-// <copyright file="local-user-config-extensions-tests.h" company="Soup">
+// <copyright file="local-user-config-extensions-tests.cpp" company="Soup">
 // Copyright (c) Soup. All rights reserved.
 // </copyright>
 
-#pragma once
+module;
+
+#include <format>
+#include <map>
+#include <memory>
+#include <set>
+#include <sstream>
+#include <string>
+#include <string_view>
+#include <vector>
+
+export module Soup.Core.Tests:LocalUserConfigExtensionsTests;
+
+import Monitor.Host;
+import Opal;
+import Soup.Core;
+import Soup.Test.Assert;
+
+using namespace Opal;
+using namespace Opal::System;
+using namespace Soup::Test;
 
 namespace Soup::Core::UnitTests
 {
-	class LocalUserConfigExtensionsTests
+	export class LocalUserConfigExtensionsTests
 	{
 	public:
 		// [[Fact]]
