+++ code/client/core/tests/operation-graph/operation-graph-manager-tests.cpp
@@ -1,12 +1,32 @@
-// <copyright file="operation-results-manager-tests.h" company="Soup">
+// <copyright file="operation-graph-manager-tests.cpp" company="Soup">
 // Copyright (c) Soup. All rights reserved.
 // </copyright>
 
-#pragma once
+module;
+
+#include <format>
+#include <map>
+#include <memory>
+#include <set>
+#include <sstream>
+#include <string>
+#include <string_view>
+#include <vector>
+
+export module Soup.Core.Tests:OperationGraphManagerTests;
+
+import Monitor.Host;
+import Opal;
+import Soup.Core;
+import Soup.Test.Assert;
+
+using namespace Opal;
+using namespace Opal::System;
+using namespace Soup::Test;
 
 namespace Soup::Core::UnitTests
 {
-	class OperationResultsManagerTests
+	export class OperationGraphManagerTests
 	{
 	public:
 		// [[Fact]]
@@ -20,17 +40,17 @@ namespace Soup::Core::UnitTests
 			auto fileSystem = std::make_shared<MockFileSystem>();
 			auto scopedFileSystem = ScopedFileSystemRegister(fileSystem);
 
-			auto filePath = Path("./TestFiles/NoFile/.soup/OperationResults.bor");
+			auto filePath = Path("./TestFiles/NoFile/.soup/OperationGraph.bog");
 			auto fileSystemState = std::make_shared<FileSystemState>();
-			auto actual = OperationResults();
-			auto result = OperationResultsManager::TryLoadState(filePath, actual, *fileSystemState);
+			auto actual = OperationGraph();
+			auto result = OperationGraphManager::TryLoadState(filePath, actual, *fileSystemState);
 
 			Assert::IsFalse(result, "Verify result is false.");
 
 			// Verify expected file system requests
 			Assert::AreEqual(
 				std::vector<std::string>({
-					"TryOpenReadBinary: ./TestFiles/NoFile/.soup/OperationResults.bor",
+					"TryOpenReadBinary: ./TestFiles/NoFile/.soup/OperationGraph.bog",
 				}),
 				fileSystem->GetRequests(),
 				"Verify file system requests match expected.");
@@ -38,7 +58,7 @@ namespace Soup::Core::UnitTests
 			// Verify expected logs
 			Assert::AreEqual(
 				std::vector<std::string>({
-					"INFO: Operation results file does not exist",
+					"INFO: Operation graph file does not exist",
 				}),
 				testListener->GetMessages(),
 				"Verify messages match expected.");
@@ -55,20 +75,20 @@ namespace Soup::Core::UnitTests
 			auto fileSystem = std::make_shared<MockFileSystem>();
 			auto scopedFileSystem = ScopedFileSystemRegister(fileSystem);
 			fileSystem->CreateMockFile(
-				Path("./TestFiles/GarbageOperationResults/.soup/OperationResults.bor"),
+				Path("./TestFiles/GarbageOperationGraph/.soup/OperationGraph.bog"),
 				std::make_shared<MockFile>(std::stringstream("garbage")));
 
-			auto filePath = Path("./TestFiles/GarbageOperationResults/.soup/OperationResults.bor");
+			auto filePath = Path("./TestFiles/GarbageOperationGraph/.soup/OperationGraph.bog");
 			auto fileSystemState = std::make_shared<FileSystemState>();
-			auto actual = OperationResults();
-			auto result = OperationResultsManager::TryLoadState(filePath, actual, *fileSystemState);
+			auto actual = OperationGraph();
+			auto result = OperationGraphManager::TryLoadState(filePath, actual, *fileSystemState);
 
 			Assert::IsFalse(result, "Verify result is false.");
 
 			// Verify expected file system requests
 			Assert::AreEqual(
 				std::vector<std::string>({
-					"TryOpenReadBinary: ./TestFiles/GarbageOperationResults/.soup/OperationResults.bor",
+					"TryOpenReadBinary: ./TestFiles/GarbageOperationGraph/.soup/OperationGraph.bog",
 				}),
 				fileSystem->GetRequests(),
 				"Verify file system requests match expected.");
@@ -76,7 +96,7 @@ namespace Soup::Core::UnitTests
 			// Verify expected logs
 			Assert::AreEqual(
 				std::vector<std::string>({
-					"ERRO: Invalid operation results file header",
+					"ERRO: Invalid operation graph file header",
 				}),
 				testListener->GetMessages(),
 				"Verify messages match expected.");
@@ -95,46 +115,69 @@ namespace Soup::Core::UnitTests
 
 			auto binaryFileContent = std::vector<uint8_t>(
 			{
-				'B', 'O', 'R', '\0', 0x02, 0x00, 0x00, 0x00,
+				'B', 'O', 'G', '\0', 0x06, 0x00, 0x00, 0x00,
 				'F', 'I', 'S', '\0', 0x00, 0x00, 0x00, 0x00,
-				'R', 'T', 'S', '\0', 0x01, 0x00, 0x00, 0x00,
+				'R', 'O', 'P', '\0', 0x01, 0x00, 0x00, 0x00,
 				0x05, 0x00, 0x00, 0x00,
+				'O', 'P', 'S', '\0', 0x01, 0x00, 0x00, 0x00,
+				0x05, 0x00, 0x00, 0x00,
+				0x0D, 0x00, 0x00, 0x00, 'T', 'e', 's', 't', 'O', 'p', 'e', 'r', 'a', 't', 'i', 'o', 'n',
+				0x08, 0x00, 0x00, 0x00, 'C', ':', '/', 'R', 'o', 'o', 't', '/',
+				0x0D, 0x00, 0x00, 0x00, '.', '/', 'D', 'o', 'S', 't', 'u', 'f', 'f', '.', 'e', 'x', 'e',
+				0x02, 0x00, 0x00, 0x00,
+				0x04, 0x00, 0x00, 0x00, 'a', 'r', 'g', '1',
+				0x04, 0x00, 0x00, 0x00, 'a', 'r', 'g', '2',
 				0x00, 0x00, 0x00, 0x00,
-				0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
 				0x00, 0x00, 0x00, 0x00,
 				0x00, 0x00, 0x00, 0x00,
+				0x00, 0x00, 0x00, 0x00,
+				0x00, 0x00, 0x00, 0x00,
+				0x01, 0x00, 0x00, 0x00,
 			});
 			fileSystem->CreateMockFile(
-				Path("./TestFiles/SimpleOperationResults/.soup/OperationResults.bor"),
+				Path("./TestFiles/SimpleOperationGraph/.soup/OperationGraph.bog"),
 				std::make_shared<MockFile>(std::stringstream(std::string((char*)binaryFileContent.data(), binaryFileContent.size()))));
 
-			auto filePath = Path("./TestFiles/SimpleOperationResults/.soup/OperationResults.bor");
+			auto filePath = Path("./TestFiles/SimpleOperationGraph/.soup/OperationGraph.bog");
 			auto fileSystemState = std::make_shared<FileSystemState>();
-			auto actual = OperationResults();
-			auto result = OperationResultsManager::TryLoadState(filePath, actual, *fileSystemState);
+			auto actual = OperationGraph();
+			auto result = OperationGraphManager::TryLoadState(filePath, actual, *fileSystemState);
 
 			Assert::IsTrue(result, "Verify result is true.");
 
-			// Verify operation Results matches expected
+			// Verify operation graph matches expected
 			Assert::AreEqual(
-				std::map<OperationId, OperationResult>({
+				std::vector<OperationId>({
+					5,
+				}),
+				actual.GetRootOperationIds(),
+				"Verify root operation ids match expected.");
+			Assert::AreEqual(
+				std::map<OperationId, OperationInfo>({
 					{
 						5,
-						OperationResult(
-							false,
-							std::chrono::clock_cast<std::chrono::file_clock>(
-								std::chrono::time_point<std::chrono::system_clock>()),
+						OperationInfo(
+							5,
+							"TestOperation",
+							CommandInfo(
+								Path("C:/Root/"),
+								Path("./DoStuff.exe"),
+								{ "arg1", "arg2" }),
 							{ },
-							{ })
-					},
+							{ },
+							{ },
+							{ },
+							{ },
+							1),
+					}
 				}),
-				actual.GetResults(),
-				"Verify results match expected.");
+				actual.GetOperations(),
+				"Verify operations match expected.");
 
 			// Verify expected file system requests
 			Assert::AreEqual(
 				std::vector<std::string>({
-					"TryOpenReadBinary: ./TestFiles/SimpleOperationResults/.soup/OperationResults.bor",
+					"TryOpenReadBinary: ./TestFiles/SimpleOperationGraph/.soup/OperationGraph.bog",
 				}),
 				fileSystem->GetRequests(),
 				"Verify file system requests match expected.");
@@ -154,24 +197,32 @@ namespace Soup::Core::UnitTests
 			auto scopedFileSystem = ScopedFileSystemRegister(fileSystem);
 
 			auto fileSystemState = std::make_shared<FileSystemState>();
-			auto filePath = Path("./TestFiles/.soup/OperationResults.bor");
-			auto operationResults = OperationResults({
-				{
+			auto filePath = Path("./TestFiles/.soup/OperationGraph.bog");
+			auto operationGraph = OperationGraph(
+				std::vector<OperationId>({
 					5,
-					OperationResult(
-						false,
-						std::chrono::clock_cast<std::chrono::file_clock>(
-							std::chrono::time_point<std::chrono::system_clock>()),
+				}),
+				std::vector<OperationInfo>({
+					OperationInfo(
+						5,
+						"TestOperation",
+						CommandInfo(
+							Path("C:/Root/"),
+							Path("./DoStuff.exe"),
+							{ "arg1", "arg2" }),
 						{ },
-						{ })
-				},
-			});
-			OperationResultsManager::SaveState(filePath, operationResults, *fileSystemState);
+						{ },
+						{ },
+						{ },
+						{ },
+						1),
+				}));
+			OperationGraphManager::SaveState(filePath, operationGraph, *fileSystemState);
 
 			// Verify expected file system requests
 			Assert::AreEqual(
 				std::vector<std::string>({
-					"OpenWriteBinary: ./TestFiles/.soup/OperationResults.bor",
+					"OpenWriteBinary: ./TestFiles/.soup/OperationGraph.bog",
 				}),
 				fileSystem->GetRequests(),
 				"Verify file system requests match expected.");
@@ -179,16 +230,26 @@ namespace Soup::Core::UnitTests
 			// Verify the file content
 			auto binaryFileContent = std::vector<uint8_t>(
 			{
-				'B', 'O', 'R', '\0', 0x02, 0x00, 0x00, 0x00,
+				'B', 'O', 'G', '\0', 0x06, 0x00, 0x00, 0x00,
 				'F', 'I', 'S', '\0', 0x00, 0x00, 0x00, 0x00,
-				'R', 'T', 'S', '\0', 0x01, 0x00, 0x00, 0x00,
+				'R', 'O', 'P', '\0', 0x01, 0x00, 0x00, 0x00,
 				0x05, 0x00, 0x00, 0x00,
+				'O', 'P', 'S', '\0', 0x01, 0x00, 0x00, 0x00,
+				0x05, 0x00, 0x00, 0x00,
+				0x0D, 0x00, 0x00, 0x00, 'T', 'e', 's', 't', 'O', 'p', 'e', 'r', 'a', 't', 'i', 'o', 'n',
+				0x08, 0x00, 0x00, 0x00, 'C', ':', '/', 'R', 'o', 'o', 't', '/',
+				0x0D, 0x00, 0x00, 0x00, '.', '/', 'D', 'o', 'S', 't', 'u', 'f', 'f', '.', 'e', 'x', 'e',
+				0x02, 0x00, 0x00, 0x00,
+				0x04, 0x00, 0x00, 0x00, 'a', 'r', 'g', '1',
+				0x04, 0x00, 0x00, 0x00, 'a', 'r', 'g', '2',
 				0x00, 0x00, 0x00, 0x00,
-				0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
 				0x00, 0x00, 0x00, 0x00,
 				0x00, 0x00, 0x00, 0x00,
+				0x00, 0x00, 0x00, 0x00,
+				0x00, 0x00, 0x00, 0x00,
+				0x01, 0x00, 0x00, 0x00,
 			});
-			auto mockFile = fileSystem->GetMockFile(Path("./TestFiles/.soup/OperationResults.bor"));
+			auto mockFile = fileSystem->GetMockFile(Path("./TestFiles/.soup/OperationGraph.bog"));
 			Assert::AreEqual(
 				std::string((char*)binaryFileContent.data(), binaryFileContent.size()),
 				mockFile->Content.str(),
