+++ code/client/core/tests/operation-graph/operation-graph-reader-tests.cpp
@@ -1,12 +1,32 @@
-// <copyright file="operation-graph-reader-tests.h" company="Soup">
+// <copyright file="operation-graph-reader-tests.cpp" company="Soup">
 // Copyright (c) Soup. All rights reserved.
 // </copyright>
 
-#pragma once
+module;
+
+#include <format>
+#include <map>
+#include <memory>
+#include <set>
+#include <sstream>
+#include <string>
+#include <string_view>
+#include <vector>
+
+export module Soup.Core.Tests:OperationGraphReaderTests;
+
+import Monitor.Host;
+import Opal;
+import Soup.Core;
+import Soup.Test.Assert;
+
+using namespace Opal;
+using namespace Opal::System;
+using namespace Soup::Test;
 
 namespace Soup::Core::UnitTests
 {
-	class OperationGraphReaderTests
+	export class OperationGraphReaderTests
 	{
 	public:
 		// [[Fact]]
