+++ code/client/core/tests/operation-graph/operation-results-reader-tests.cpp
@@ -1,13 +1,35 @@
-// <copyright file="operation-results-reader-tests.h" company="Soup">
+// <copyright file="operation-results-reader-tests.cpp" company="Soup">
 // Copyright (c) Soup. All rights reserved.
 // </copyright>
 
-#pragma once
+module;
+
+#include <chrono>
+#include <format>
+#include <map>
+#include <memory>
+#include <set>
+#include <sstream>
+#include <string>
+#include <string_view>
+#include <vector>
+
+export module Soup.Core.Tests:OperationResultsReaderTests;
+
+import Monitor.Host;
+import Opal;
+import Soup.Core;
+import Soup.Test.Assert;
+
+using namespace Opal;
+using namespace Opal::System;
+using namespace Soup::Test;
 using namespace std::chrono;
+using namespace std::chrono_literals;
 
 namespace Soup::Core::UnitTests
 {
-	class OperationResultsReaderTests
+	export class OperationResultsReaderTests
 	{
 	public:
 		// [[Fact]]
